Shift playback to middle of screen, add playback line
