+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/narration/Narration.kt
@@ -318,7 +318,10 @@ class Narration @AssistedInject constructor(
 
     fun scrollAudio(delta: Int) {
         chapterReaderConnection.seek(delta)
-        player
+    }
+
+    fun seek(location: Int) {
+        chapterReaderConnection.seek(location)
     }
 
     fun getLocationInFrames(): Int {
